telegram-bot/requirements.txt
python-telegram-bot==21.10
websockets==14.2
